Disable the reports weight calculation
This was supposed to be used for automatic action taken based on the
report weight. We do not use it that way though. Additionally, the
current implementation was strictly bound to moodle.org use case. With
the hope there will be some better general metric available one day, we
are simply setting the weight to 1 for now.

Credit goes to Dan for this easy to implement suggestion.